# Удалён неиспользуемый класс ContactsVM
