+++ src/ContactsApp/View/Model/Contact.cs
@@ -10,27 +10,74 @@ namespace View.Model
     /// <summary>
     /// Представляет контактную информацию.
     /// </summary>
-    public class Contact
+    public class Contact : INotifyPropertyChanged, IDataErrorInfo
     {
+        private string _name;
+        private string _phone;
+        private string _email;
+
         /// <summary>
         /// Получает или задаёт ФИО контакта.
         /// </summary>
-        public string Name { get; set; }
+        public string Name
+        {
+            get => _name;
+            set
+            {
+                _name = value;
+                OnPropertyChanged(nameof(Name));
+            }
+        }
 
         /// <summary>
         /// Получает или задаёт номер телефона контакта.
         /// </summary>
-        public string Phone { get; set; }
+        public string Phone
+        {
+            get => _phone;
+            set
+            {
+                _phone = value;
+                OnPropertyChanged(nameof(Phone));
+            }
+        }
 
         /// <summary>
         /// Получает или задаёт адрес электронной почты контакта.
         /// </summary>
-        public string Email { get; set; }
+        public string Email
+        {
+            get => _email;
+            set
+            {
+                _email = value;
+                OnPropertyChanged(nameof(Email));
+            }
+        }
 
         /// <summary>
         /// Инициализирует новый экземпляр класса <see cref="Contact"/>.
         /// (Конструктор по умолчанию)
         /// </summary>
         public Contact() { }
+
+        public string this[string columnName] => throw new NotImplementedException();
+
+        public string Error => throw new NotImplementedException();
+
+        /// <summary>
+        /// Возникает при изменении значения свойства.
+        /// </summary>
+        public event PropertyChangedEventHandler? PropertyChanged;
+
+        /// <summary>
+        /// Вызывает событие <see cref="PropertyChanged"/>
+        /// для уведомления об изменении свойства.
+        /// </summary>
+        /// <param name="propertyName">Имя изменённого свойства.</param>
+        private void OnPropertyChanged(string propertyName)
+        {
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+        }
     }
 }
